chore: clean up utils
